Handle another overflow case

+++ src/libraries/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableArray_1.Builder.cs
@@ -799,7 +799,7 @@ namespace System.Collections.Immutable
                 }
 
                 Requires.Range(startIndex >= 0 && startIndex <= this.Count, nameof(startIndex));
-                Requires.Range(count >= 0 && startIndex + count <= this.Count, nameof(count));
+                Requires.Range(count >= 0 && (uint)(startIndex + count) <= (uint)this.Count, nameof(count));
 
                 equalityComparer ??= EqualityComparer<T>.Default;
                 if (equalityComparer == EqualityComparer<T>.Default)
